requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.5
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
